Refactoring main function from ImageView as an example
